Modified logic for to add recipes to favorite in project integrator of module 4.

+++ module4-node-js/3-integrator-project/project/api/src/service/RecipeFavoriteService.ts
@@ -63,7 +63,7 @@ export default class RecipeFavoriteService {
 				newRecipeFavorite.userId,
 				newRecipeFavorite.recipeId
 			)) as Recipe;
-			if (recipeFavoriteWithDetails != null) throw new CustomException("Recipe favorite already exist", StatusCodes.NOT_FOUND);
+			if (recipeFavoriteWithDetails != null) throw new CustomException("Recipe favorite already exist", StatusCodes.CONFLICT);
 			const recipeFavoriteCreated: RecipeFavorite = await RecipeFavoriteRepository.createRecipeFavorite(newRecipeFavorite);
 			return recipeFavoriteCreated;
 		} catch (error: any) {
@@ -74,9 +74,11 @@ export default class RecipeFavoriteService {
 	static deleteRecipeFavoriteByUserIdAndRecipeId = async (userId: number, recipeId: number): Promise<any> => {
 		try {
 			const recipeFavoriteWithDetails: Recipe = await this.getRecipeFavoriteWithDetailsByUserIdAndRecipeId(userId, recipeId);
-			const response = await RecipeFavoriteRepository.deleteRecipeFavoriteByUserIdAndRecipeId(userId, recipeId);
-			if (response.affected > 0) {
-				return recipeFavoriteWithDetails;
+			if (recipeFavoriteWithDetails != null) {
+				const response = await RecipeFavoriteRepository.deleteRecipeFavoriteByUserIdAndRecipeId(userId, recipeId);
+				if (response.affected > 0) {
+					return recipeFavoriteWithDetails;
+				}
 			}
 			return null;
 		} catch (error: any) {
